requirements.txt
numpy==1.23.5
requests>=2.31.0
farm-haystack[sql]==1.15.0
faiss-cpu==1.7.4
streamlit==1.24.0
protobuf<=3.20.2
pydantic<2.0.0
